endpoints/requirements.txt
alembic==1.11.1
apifairy==1.3.0
apispec==6.3.0
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
fonttools==4.39.4
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
typing-inspect==0.9.0
typing_extensions==4.6.2
webargs==8.2.0
Werkzeug==2.3.4
